[major] Refactor the Executor interface
Replace the backend parameter from the initialization of the Executor class with individual executor classes for the individual backends:
* `Executor(backend="local")` becomes `from executorlib import LocalExecutor§
* `Executor(backend="flux_allocation")` becomes `from executorlib import FluxAllocationExecutor§
* `Executor(backend="flux_submission")` becomes `from executorlib import FluxSubmissionExecutor§
* `Executor(backend="slurm_allocation")` becomes "SlurmAllocationExecutor"
* `Executor(backend="slurm_submission")` becomes `from executorlib import SlurmSubmissionExecutor`

This has two advantages: On the one hand it is less error prone to mistyping the backend name, as the user can use auto completion to import the right module. On the other hand it is more consistent with the standard library which defines the `ProcessPoolExecutor` and the `ThreadPoolExecutor`, rather than a `backend` parameter.
