+++ tests/test_executor_backend_flux.py
@@ -3,7 +3,7 @@ import unittest
 
 import numpy as np
 
-from executorlib import Executor
+from executorlib import FluxAllocationExecutor
 
 
 try:
@@ -44,10 +44,9 @@ class TestFluxBackend(unittest.TestCase):
         self.executor = flux.job.FluxExecutor()
 
     def test_flux_executor_serial(self):
-        with Executor(
+        with FluxAllocationExecutor(
             max_cores=2,
             flux_executor=self.executor,
-            backend="flux_allocation",
             block_allocation=True,
         ) as exe:
             fs_1 = exe.submit(calc, 1)
@@ -58,11 +57,10 @@ class TestFluxBackend(unittest.TestCase):
             self.assertTrue(fs_2.done())
 
     def test_flux_executor_threads(self):
-        with Executor(
+        with FluxAllocationExecutor(
             max_cores=1,
             resource_dict={"threads_per_core": 2},
             flux_executor=self.executor,
-            backend="flux_allocation",
             block_allocation=True,
         ) as exe:
             fs_1 = exe.submit(calc, 1)
@@ -73,11 +71,10 @@ class TestFluxBackend(unittest.TestCase):
             self.assertTrue(fs_2.done())
 
     def test_flux_executor_parallel(self):
-        with Executor(
+        with FluxAllocationExecutor(
             max_cores=2,
             resource_dict={"cores": 2},
             flux_executor=self.executor,
-            backend="flux_allocation",
             block_allocation=True,
             flux_executor_pmi_mode=pmi,
         ) as exe:
@@ -86,11 +83,10 @@ class TestFluxBackend(unittest.TestCase):
             self.assertTrue(fs_1.done())
 
     def test_single_task(self):
-        with Executor(
+        with FluxAllocationExecutor(
             max_cores=2,
             resource_dict={"cores": 2},
             flux_executor=self.executor,
-            backend="flux_allocation",
             block_allocation=True,
             flux_executor_pmi_mode=pmi,
         ) as p:
@@ -105,11 +101,10 @@ class TestFluxBackend(unittest.TestCase):
         os.makedirs(dirname, exist_ok=True)
         file_stdout = os.path.join(dirname, "flux.out")
         file_stderr = os.path.join(dirname, "flux.err")
-        with Executor(
+        with FluxAllocationExecutor(
             max_cores=1,
             resource_dict={"cores": 1, "cwd": dirname},
             flux_executor=self.executor,
-            backend="flux_allocation",
             block_allocation=True,
             flux_log_files=True,
         ) as p:
@@ -126,11 +121,10 @@ class TestFluxBackend(unittest.TestCase):
     def test_output_files_abs(self):
         file_stdout = os.path.abspath("flux.out")
         file_stderr = os.path.abspath("flux.err")
-        with Executor(
+        with FluxAllocationExecutor(
             max_cores=1,
             resource_dict={"cores": 1},
             flux_executor=self.executor,
-            backend="flux_allocation",
             block_allocation=True,
             flux_log_files=True,
         ) as p:
@@ -145,12 +139,11 @@ class TestFluxBackend(unittest.TestCase):
         os.remove(file_stderr)
 
     def test_internal_memory(self):
-        with Executor(
+        with FluxAllocationExecutor(
             max_cores=1,
             resource_dict={"cores": 1},
             init_function=set_global,
             flux_executor=self.executor,
-            backend="flux_allocation",
             block_allocation=True,
         ) as p:
             f = p.submit(get_global)
@@ -160,10 +153,9 @@ class TestFluxBackend(unittest.TestCase):
 
     def test_validate_max_workers(self):
         with self.assertRaises(ValueError):
-            Executor(
+            FluxAllocationExecutor(
                 max_workers=10,
                 resource_dict={"cores": 10, "threads_per_core": 10},
                 flux_executor=self.executor,
-                backend="flux_allocation",
                 block_allocation=True,
             )
