+++ tests/test_integration_pyiron_workflow.py
@@ -12,7 +12,7 @@ from time import sleep
 from typing import Callable
 import unittest
 
-from executorlib import Executor
+from executorlib import LocalExecutor
 from executorlib.standalone.serialize import cloudpickle_register
 
 
@@ -75,7 +75,7 @@ class TestDynamicallyDefinedObjects(unittest.TestCase):
             return dynamic_arg
 
         dynamic_dynamic = slowly_returns_dynamic()
-        executor = Executor(block_allocation=True, max_workers=1)
+        executor = LocalExecutor(block_allocation=True, max_workers=1)
         cloudpickle_register(ind=1)
         dynamic_object = does_nothing()
         fs = executor.submit(dynamic_dynamic.run, dynamic_object)
@@ -105,7 +105,7 @@ class TestDynamicallyDefinedObjects(unittest.TestCase):
         self.assertIsNone(
             dynamic_42.result, msg="Just a sanity check that the test is set up right"
         )
-        executor = Executor(block_allocation=True, max_workers=1)
+        executor = LocalExecutor(block_allocation=True, max_workers=1)
         cloudpickle_register(ind=1)
         fs = executor.submit(dynamic_42.run)
         fs.add_done_callback(dynamic_42.process_result)
@@ -136,7 +136,7 @@ class TestDynamicallyDefinedObjects(unittest.TestCase):
             dynamic_42.running,
             msg="Sanity check that the test starts in the expected condition",
         )
-        executor = Executor(block_allocation=True, max_workers=1)
+        executor = LocalExecutor(block_allocation=True, max_workers=1)
         cloudpickle_register(ind=1)
         fs = executor.submit(dynamic_42.run)
         fs.add_done_callback(dynamic_42.process_result)
@@ -160,7 +160,7 @@ class TestDynamicallyDefinedObjects(unittest.TestCase):
             raise RuntimeError
 
         re = raise_error()
-        executor = Executor(block_allocation=True, max_workers=1)
+        executor = LocalExecutor(block_allocation=True, max_workers=1)
         cloudpickle_register(ind=1)
         fs = executor.submit(re.run)
         with self.assertRaises(
@@ -190,7 +190,7 @@ class TestDynamicallyDefinedObjects(unittest.TestCase):
             return inside_variable
 
         dynamic_dynamic = slowly_returns_dynamic()
-        executor = Executor(block_allocation=True, max_workers=1)
+        executor = LocalExecutor(block_allocation=True, max_workers=1)
         cloudpickle_register(ind=1)
         fs = executor.submit(dynamic_dynamic.run)
         self.assertIsInstance(
@@ -219,7 +219,7 @@ class TestDynamicallyDefinedObjects(unittest.TestCase):
             return fortytwo
 
         f = slow()
-        executor = Executor(block_allocation=True, max_workers=1)
+        executor = LocalExecutor(block_allocation=True, max_workers=1)
         cloudpickle_register(ind=1)
         fs = executor.submit(f.run)
         self.assertEqual(
